Autorization checking and @can declarations for client managing

+++ app/Http/Controllers/ClientController.php
@@ -25,9 +25,12 @@ class ClientController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create');
+
         $title = 'Client create';
 
         return view('client.create', compact('title'));
@@ -38,9 +41,12 @@ class ClientController extends Controller
      *
      * @param  CreateUpdateClientRequest  $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(CreateUpdateClientRequest $request)
     {
+        $this->authorize('create');
+
         Client::create($request->validated());
 
         return redirect(route('client.index'))->with('created', true);;
@@ -62,9 +68,12 @@ class ClientController extends Controller
      *
      * @param  Client  $client
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Client $client)
     {
+        $this->authorize('update', $client);
+
         $title = 'Client edit: '.$client->company;
 
         return view('client.edit', compact('title', 'client'));
@@ -76,9 +85,12 @@ class ClientController extends Controller
      * @param  CreateUpdateClientRequest  $request
      * @param  Client  $client
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(CreateUpdateClientRequest $request, Client $client)
     {
+        $this->authorize('update', $client);
+
         $client->update($request->validated());
 
         return redirect(route('client.edit', $client->id))->with('updated', true);
@@ -89,9 +101,12 @@ class ClientController extends Controller
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete', $id);
+
         Client::destroy([$id]);
 
         return redirect(route('client.index'))->with('deleted', true);
@@ -116,9 +131,12 @@ class ClientController extends Controller
      *
      * @param $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function restore($id)
     {
+        $this->authorize('restore', $id);
+
         Client::withTrashed()->where('id', $id)->restore();
 
         return redirect(route('client.deleted'))->with('restored', true);
